requirements.txt
pywin32; sys_platform == "win32"
python-xlib; sys_platform == "linux"
pypresence
